refactor(desktop): split AppState coordination
